backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pandas>=2.0.0
openpyxl==3.1.2
google-generativeai==0.7.2
slowapi==0.1.9
python-dotenv==1.0.1
aiosmtplib==3.0.1
email-validator==2.1.1
httpx==0.27.0